requirements.txt
Flask>=3.0.0,<4.0.0
pandas>=2.1.0,<2.3.0
openpyxl>=3.1.0,<3.2.0
gunicorn>=21.0.0,<22.0.0
azure-storage-blob>=12.19.0,<13.0.0